restore stan.py; move previous stan.py to utils.py
